add tb docs on dynamic tools
